fix(differentDomains): Match domain properly on client if port provided
If locale's "domain" contained a port then that wasn't matched correctly
on the client as it compared with "location.hostname" that doesn't include
the port. Use "location.host" instead.

Also, force "no-prefix" strategy when using "differentDomains" as
otherwise duplicate routes are created with the same paths but different
names. That has then caused issues when detecting locale from the route
as it wouldn't detect different locale and override what was set on
"$i18n.locale" previously.

Resolves #831